consolidate plotting helpers
